Switch types from int to Py_ssize_t to be compatible with Python 2.5 on 64-bit archs

+++ _dbus_bindings/conn-methods.c
@@ -141,7 +141,7 @@ _filter_message(DBusConnection *conn, DBusMessage *message, void *user_data)
     PyObject *callable = NULL;
     PyObject *msg_obj;
 #ifndef DBUS_PYTHON_DISABLE_CHECKS
-    int i, size;
+    Py_ssize_t i, size;
 #endif
 
     dbus_message_ref(message);
